+++ src/main/java/de/soderer/utilities/json/schema/validator/MaxPropertiesValidator.java
@@ -31,7 +31,7 @@ public class MaxPropertiesValidator extends BaseJsonSchemaValidator {
 			}
 		} else {
 			if (((JsonObject) jsonNode.getValue()).keySet().size() > ((Integer) validatorData)) {
-				throw new JsonSchemaDataValidationError("Required maximum number of properties is '" + ((Integer) validatorData) + "' but was '" + ((JsonObject) jsonNode.getValue()).keySet().size()  + "'", jsonPath);
+				throw new JsonSchemaDataValidationError("Required maximum number of properties is '" + validatorData + "' but was '" + ((JsonObject) jsonNode.getValue()).keySet().size()  + "'", jsonPath);
 			}
 		}
     }
