+++ src/main/java/de/soderer/utilities/json/schema/validator/MinLengthValidator.java
@@ -30,7 +30,7 @@ public class MinLengthValidator extends BaseJsonSchemaValidator {
 			}
 		} else {
 			if (((String) jsonNode.getValue()).length() < ((Integer) validatorData)) {
-				throw new JsonSchemaDataValidationError("String minLength is '" + ((Integer) validatorData) + "' but was '" + ((String) jsonNode.getValue()).length()  + "'", jsonPath);
+				throw new JsonSchemaDataValidationError("String minLength is '" + validatorData + "' but was '" + ((String) jsonNode.getValue()).length()  + "'", jsonPath);
 			}
 		}
     }
